FileClass.java added new.txt added renamed folder to new_folder
