+++ src/com/some_folder/FileClass.java
@@ -1,0 +1,4 @@
+package com.some_folder;
+
+public class FileClass {
+}
