+++ src/com/some_folder/vv_input.txt
@@ -1,6 +1,7 @@
 line 1
 line 2
 line 3
+
 line 4
 line 5
-added something
+added someing line
