+++ src/com/some_folder/vvv_input.txt
